Fixes and extract heading button components
